Improved Light Mode
Light Mode is no longer the equivalent of staring into the sun. I am unsatisfied with this refactor and will likely revisit before deployment. I actually scrapped a lot of light mode stuff in favor of resuing dark mode and while it looks better, there is more to be desired.