Simplified Backend
Removed the game folder and added the files that were in it to utils.

Refactored the routes and respective handlers so multiple requests are no longer made at the same time. There is now a round request every time the user presses Submit Bets and a game request at the end. All the logic inside of the requests is the same, just combined and simplified.