Added Spins, Games Completed, Total Won
Swapped RoundStats with AllStats, as it now includes match stats. If I decide to track user betting stats, I will likyl make this RoundStats again and then create a MatchStats or UserStats model. Fixed the issue where buttons that only have borders in light mode flash on transition. Tinkered with the spacing of the footer links, now when a user copies the share link, the others don't move.