Improved Screen Responsiveness
This will be a never-ending battle, but I think I made good progress here. I added a clamp to fix mobile responsiveness. When width drops below 1024, it will use the clamp. Media queries are still used for anything larger, so it is likely that there will still be some issues there.