Fixed Off by 1 Bugs
There was some bugs in the payout logic due to 3 having a grid index of 0 and not returning early whenever possible. Now, if the user bets on an exact number and doesn't hit, it returns right immediately. Additionally, the inner bet logic is not called if the result number is 0 or 00.